patch/dummy_enclosure
